Add basiv functionality to allow the top nav bar, to better handle whether a user is logged in or out.
Fixed a bug in the localStorageProvider remove method. I did not have the required breaks in teh switch statement, so it was falling through.

Changed the logout method in the AuthenticationProvider to now also return a promise.

Also added a event aggregator to the AuthenticationProvider, so that it broadcasts a simple true or false if the login state changes.